// src/types/sale.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SalesPulse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from './role';
import type { SaleItem } from './saleItem';
import type { SalePayment } from './salePayment';
import type { SaleStatus } from './saleStatus';

export interface Sale {
  id: number;
  createdAt: Date;
  operatorRole: Role;
  status: SaleStatus;
  items: SaleItem[];
  payments: SalePayment[];
  subtotalCents: number;
  discountCents: number;
  totalCents: number;
  /** @nullable */
  customerId: number | null;
  /** @nullable */
  customerName: string | null;
  /** @nullable */
  voidReason: string | null;
  /** @nullable */
  voidedAt: Date | null;
}